backend/requirements.txt
fastapi
uvicorn
tiktoken
ollama
python-dotenv
